// x-pack/platform/packages/shared/agent-builder/agent-builder-server/hooks/types.d.ts
interface AgentHookContextBase {
    request: KibanaRequest;
    abortSignal?: AbortSignal;
    agentId?: string;
}
export interface BeforeAgentHookContext extends AgentHookContextBase {
    nextInput: ProcessedRoundInput;
}
interface ToolCallHookContextBase extends AgentHookContextBase {
    toolId: string;
    toolCallId: string;
    toolParams: Record<string, unknown>;
    source: ToolCallSource;
}
export type BeforeToolCallHookContext = ToolCallHookContextBase;
export interface AfterToolCallHookContext extends ToolCallHookContextBase {
    toolReturn: RunToolReturn;
    toolHandlerContext: ToolHandlerContext;
}