// x-pack/platform/packages/shared/agent-builder/agent-builder-server/runner/runner.d.ts
export interface RunToolReturn {
    /**
     * The result value as returned by the tool.
     */
    results?: ToolResult[];
    /**
     * Interruption returned by the tool
     */
    prompt?: PromptRequest;
}